Add Hero page
